Remove duplicate browser-run helper ownership
