+++ test/agent-browser.extension-click-dispatch.test.ts
@@ -1,0 +1,124 @@
+/**
+ * Purpose: Verify click-dispatch probe diagnostics and lifecycle cleanup for the pi-agent-browser extension.
+ * Responsibilities: Assert click-dispatch probes report success-only diagnostics and clean up after failed clicks.
+ * Scope: Integration-style Node test-runner coverage with fake agent-browser binaries.
+ */
+
+import assert from "node:assert/strict";
+import { mkdtemp, rm } from "node:fs/promises";
+import { tmpdir } from "node:os";
+import { join } from "node:path";
+import test from "node:test";
+
+import {
+	createExtensionHarness,
+	executeRegisteredTool,
+	readInvocationLog,
+	runExtensionEvent,
+	withPatchedEnv,
+	writeFakeAgentBrowserBinary,
+} from "./helpers/agent-browser-harness.js";
+
+test("agentBrowserExtension cleans up click dispatch probes after failed clicks", { concurrency: false }, async () => {
+	const tempDir = await mkdtemp(join(tmpdir(), "pi-agent-browser-click-dispatch-failure-"));
+	const logPath = join(tempDir, "invocations.log");
+	const basePath = process.env.PATH ?? "";
+	await writeFakeAgentBrowserBinary(
+		tempDir,
+		`const fs = require("node:fs");
+const args = process.argv.slice(2);
+const stdin = fs.readFileSync(0, "utf8");
+fs.appendFileSync(${JSON.stringify(logPath)}, JSON.stringify({ args, stdin }) + "\\n");
+if (args.includes("eval")) {
+  if (stdin.includes("window[marker] = state")) {
+    process.stdout.write(JSON.stringify({ success: true, data: { result: { status: "installed" } } }));
+  } else if (stdin.includes("cleaned-up")) {
+    process.stdout.write(JSON.stringify({ success: true, data: { result: { status: "cleaned-up" } } }));
+  } else {
+    process.stdout.write(JSON.stringify({ success: true, data: { result: { status: "unexpected" } } }));
+  }
+} else if (args.includes("click")) {
+  process.stdout.write(JSON.stringify({ success: false, error: "click failed" }));
+  process.exit(2);
+} else {
+  process.stdout.write(JSON.stringify({ success: true, data: "ok" }));
+}`,
+	);
+
+	try {
+		await withPatchedEnv({ PATH: `${tempDir}:${basePath}` }, async () => {
+			const harness = createExtensionHarness({ cwd: tempDir });
+			await runExtensionEvent(harness.handlers, "session_start", { reason: "new" }, harness.ctx);
+
+			const click = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["click", "[data-test=save]"] });
+			assert.equal(click.isError, true);
+			assert.equal(click.details?.clickDispatch, undefined);
+
+			const invocations = await readInvocationLog(logPath);
+			const evalInvocations = invocations.filter((entry) => entry.args.includes("eval"));
+			assert.equal(invocations.filter((entry) => entry.args.includes("click")).length, 1);
+			assert.ok(evalInvocations.some((entry) => (entry.stdin ?? "").includes("window[marker] = state")));
+			assert.ok(evalInvocations.some((entry) => (entry.stdin ?? "").includes("cleaned-up")));
+			assert.equal(evalInvocations.some((entry) => (entry.stdin ?? "").includes("no-native-event-observed")), false);
+		});
+	} finally {
+		await rm(tempDir, { force: true, recursive: true });
+	}
+});
+
+test("agentBrowserExtension reports click dispatch diagnostic when upstream reports success without dispatching DOM events", { concurrency: false }, async () => {
+	const tempDir = await mkdtemp(join(tmpdir(), "pi-agent-browser-click-dispatch-"));
+	const logPath = join(tempDir, "invocations.log");
+	const basePath = process.env.PATH ?? "";
+	await writeFakeAgentBrowserBinary(
+		tempDir,
+		`const fs = require("node:fs");
+const args = process.argv.slice(2);
+const stdin = fs.readFileSync(0, "utf8");
+fs.appendFileSync(${JSON.stringify(logPath)}, JSON.stringify({ args, stdin }) + "\\n");
+if (args.includes("snapshot")) {
+  process.stdout.write(JSON.stringify({ success: true, data: {
+    origin: "https://shop.example/inventory",
+    refs: {
+      e1: { role: "button", name: "Add to cart" },
+      e2: { role: "button", name: "Add to cart" }
+    },
+    snapshot: '- button "Add to cart" [ref=e1]\\n- button "Add to cart" [ref=e2]'
+  } }));
+} else if (args.includes("click")) {
+  process.stdout.write(JSON.stringify({ success: true, data: { clicked: args[args.length - 1] } }));
+} else if (args.includes("eval")) {
+  if (stdin.includes("window[marker] = state")) {
+    process.stdout.write(JSON.stringify({ success: true, data: { result: { status: "installed" } } }));
+  } else if (stdin.includes("no-native-event-observed")) {
+    process.stdout.write(JSON.stringify({ success: true, data: { result: { status: "no-native-event-observed", nativeEventCount: 0 } } }));
+  } else {
+    process.stdout.write(JSON.stringify({ success: true, data: { result: { title: "Shop", url: "https://shop.example/inventory" } } }));
+  }
+} else {
+  process.stdout.write(JSON.stringify({ success: true, data: "ok" }));
+}`,
+	);
+
+	try {
+		await withPatchedEnv({ PATH: `${tempDir}:${basePath}` }, async () => {
+			const harness = createExtensionHarness({ cwd: tempDir });
+			await runExtensionEvent(harness.handlers, "session_start", { reason: "new" }, harness.ctx);
+
+			const click = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["click", "[data-test=add-to-cart]"] });
+			assert.equal(click.isError, true);
+			assert.match((click.content[0] as { text: string }).text, /Click dispatch diagnostic:/);
+			assert.equal((click.details?.clickDispatch as { status?: string } | undefined)?.status, "no-native-event-observed");
+			assert.equal((click.details?.clickDispatch as { target?: { kind?: string; selector?: string } } | undefined)?.target?.kind, "selector");
+			assert.equal((click.details?.clickDispatch as { target?: { kind?: string; selector?: string } } | undefined)?.target?.selector, "[data-test=add-to-cart]");
+			assert.ok((click.details?.nextActions as Array<{ id?: string }> | undefined)?.some((action) => action.id === "retry-click-after-dispatch-miss"));
+
+			const invocations = await readInvocationLog(logPath);
+			assert.equal(invocations.filter((entry) => entry.args.includes("click")).length, 1);
+			assert.ok(invocations.some((entry) => entry.args.includes("eval") && (entry.stdin ?? "").includes("window[marker] = state")));
+			assert.ok(invocations.some((entry) => entry.args.includes("eval") && (entry.stdin ?? "").includes("no-native-event-observed")));
+		});
+	} finally {
+		await rm(tempDir, { force: true, recursive: true });
+	}
+});
