+++ test/agent-browser.extension-errors-artifacts.test.ts
@@ -7,7 +7,7 @@
  */
 
 import assert from "node:assert/strict";
-import { mkdir, mkdtemp, readFile, rm, stat, writeFile } from "node:fs/promises";
+import { mkdir, mkdtemp, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import test from "node:test";
@@ -21,7 +21,6 @@ import {
 	createSecureTempDirectory,
 } from "../extensions/agent-browser/lib/temp.js";
 import {
-	TEST_SESSION_ID,
 	createExtensionHarness,
 	createToolBranchEntry,
 	executeRegisteredTool,
@@ -710,251 +709,6 @@ if (args.includes("open")) {
 			const closeCandidate = await executeRegisteredTool(harness.tool, harness.ctx, { args: closeCandidateArgs });
 			assert.equal(closeCandidate.isError, false);
 			assert.notEqual(closeCandidate.details?.failureCategory, "stale-ref");
-		});
-	} finally {
-		await rm(tempDir, { force: true, recursive: true });
-	}
-});
-
-test("agentBrowserExtension blocks likely final order clicks when the user set a stop boundary", { concurrency: false }, async () => {
-	const tempDir = await mkdtemp(join(tmpdir(), "pi-agent-browser-stop-boundary-"));
-	const logPath = join(tempDir, "invocations.log");
-	const basePath = process.env.PATH ?? "";
-	await writeFakeAgentBrowserBinary(
-		tempDir,
-		`const fs = require("node:fs");
-const args = process.argv.slice(2);
-const stdin = fs.readFileSync(0, "utf8");
-fs.appendFileSync(${JSON.stringify(logPath)}, JSON.stringify({ args, stdin }) + "\\n");
-if (args.includes("snapshot")) {
-  process.stdout.write(JSON.stringify({ success: true, data: { origin: "https://shop.example/checkout-step-two", refs: { e9: { role: "button", name: "Finish" } }, snapshot: '- button "Finish" [ref=e9]' } }));
-} else if (args.includes("click") || args.includes("batch")) {
-  process.stdout.write(JSON.stringify({ success: true, data: { clicked: args[args.length - 1] } }));
-} else {
-  process.stdout.write(JSON.stringify({ success: true, data: "ok" }));
-}`,
-	);
-
-	try {
-		await withPatchedEnv({ PATH: `${tempDir}:${basePath}` }, async () => {
-			const harness = createExtensionHarness({ cwd: tempDir, prompt: "Stop on the checkout overview page; do not place the order." });
-			await runExtensionEvent(harness.handlers, "session_start", { reason: "new" }, harness.ctx);
-
-			const snapshot = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["snapshot", "-i"] });
-			assert.equal(snapshot.isError, false, JSON.stringify(snapshot));
-
-			const refClick = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["click", "@e9"] });
-			assert.equal(refClick.isError, true);
-			assert.match((refClick.content[0] as { text: string }).text, /Blocked likely final submit\/order click-like action/);
-			assert.equal((refClick.details?.promptGuard as { reason?: string } | undefined)?.reason, "explicit-user-stop-boundary");
-
-			const batchClick = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["batch"], stdin: JSON.stringify([["click", "[data-test=finish]"]]) });
-			assert.equal(batchClick.isError, true);
-			assert.equal((batchClick.details?.promptGuard as { stepIndex?: number } | undefined)?.stepIndex, 0);
-
-			const invocations = await readInvocationLog(logPath);
-			assert.equal(invocations.filter((entry) => entry.args.includes("click") || entry.args.includes("batch")).length, 0);
-		});
-	} finally {
-		await rm(tempDir, { force: true, recursive: true });
-	}
-});
-
-test("agentBrowserExtension blocks Enter/Return key submits but allows keyboard type text when the user set a stop boundary", { concurrency: false }, async () => {
-	const tempDir = await mkdtemp(join(tmpdir(), "pi-agent-browser-stop-boundary-keyboard-"));
-	const logPath = join(tempDir, "invocations.log");
-	const basePath = process.env.PATH ?? "";
-	await writeFakeAgentBrowserBinary(
-		tempDir,
-		`const fs = require("node:fs");
-const args = process.argv.slice(2);
-fs.appendFileSync(${JSON.stringify(logPath)}, JSON.stringify({ args }) + "\\n");
-process.stdout.write(JSON.stringify({ success: true, data: "ok" }));`,
-	);
-
-	try {
-		await withPatchedEnv({ PATH: `${tempDir}:${basePath}` }, async () => {
-			const harness = createExtensionHarness({ cwd: tempDir, prompt: "Stop on the checkout overview page; do not place the order." });
-			await runExtensionEvent(harness.handlers, "session_start", { reason: "new" }, harness.ctx);
-
-			const pressEnter = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["press", "Enter"] });
-			assert.equal(pressEnter.isError, true);
-			assert.match((pressEnter.content[0] as { text: string }).text, /keyboard submit \(Enter\/Return\)/);
-			assert.equal((pressEnter.details?.promptGuard as { reason?: string } | undefined)?.reason, "explicit-user-stop-boundary");
-
-			const keyReturn = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["key", "Return"] });
-			assert.equal(keyReturn.isError, true);
-			assert.equal((keyReturn.details?.promptGuard as { reason?: string } | undefined)?.reason, "explicit-user-stop-boundary");
-
-			const keyboardTypeEnter = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["keyboard", "type", "Enter"] });
-			assert.equal(keyboardTypeEnter.isError, false);
-
-			const batchEnter = await executeRegisteredTool(harness.tool, harness.ctx, {
-				args: ["batch"],
-				stdin: JSON.stringify([["fill", "#notes", "done"], ["press", "Enter"]]),
-			});
-			assert.equal(batchEnter.isError, true);
-			assert.equal((batchEnter.details?.promptGuard as { stepIndex?: number } | undefined)?.stepIndex, 1);
-
-			const batchReturn = await executeRegisteredTool(harness.tool, harness.ctx, {
-				args: ["batch"],
-				stdin: JSON.stringify([["fill", "#notes", "done"], ["key", "Return"]]),
-			});
-			assert.equal(batchReturn.isError, true);
-			assert.equal((batchReturn.details?.promptGuard as { stepIndex?: number } | undefined)?.stepIndex, 1);
-
-			const invocations = await readInvocationLog(logPath);
-			assert.equal(invocations.length, 1);
-			assert.deepEqual(invocations[0]?.args.slice(-3), ["keyboard", "type", "Enter"]);
-		});
-	} finally {
-		await rm(tempDir, { force: true, recursive: true });
-	}
-});
-
-test("agentBrowserExtension blocks close until required prompt screenshot artifacts are saved", { concurrency: false }, async () => {
-	const tempDir = await mkdtemp(join(tmpdir(), "pi-agent-browser-required-artifact-"));
-	const logPath = join(tempDir, "invocations.log");
-	const screenshotPath = join(tempDir, "release-smoke.png");
-	const basePath = process.env.PATH ?? "";
-	await writeFakeAgentBrowserBinary(
-		tempDir,
-		`const fs = require("node:fs");
-const args = process.argv.slice(2);
-fs.appendFileSync(${JSON.stringify(logPath)}, JSON.stringify({ args }) + "\\n");
-if (args.includes("screenshot")) {
-  const path = args[args.length - 1];
-  fs.writeFileSync(path, "png");
-  process.stdout.write(JSON.stringify({ success: true, data: { path } }));
-} else if (args.includes("close")) {
-  process.stdout.write(JSON.stringify({ success: true, data: { closed: true } }));
-} else {
-  process.stdout.write(JSON.stringify({ success: true, data: "ok" }));
-}`,
-	);
-
-	try {
-		await withPatchedEnv({ PATH: `${tempDir}:${basePath}` }, async () => {
-			const harness = createExtensionHarness({ cwd: tempDir, prompt: `Save a screenshot here: ${screenshotPath}` });
-			await runExtensionEvent(harness.handlers, "session_start", { reason: "new" }, harness.ctx);
-
-			const blockedClose = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["close"] });
-			assert.equal(blockedClose.isError, true);
-			assert.match((blockedClose.content[0] as { text: string }).text, /requested artifact path is missing or unverified/);
-			assert.equal((blockedClose.details?.promptGuard as { missingArtifacts?: Array<{ path?: string }> } | undefined)?.missingArtifacts?.[0]?.path, screenshotPath);
-
-			const screenshot = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["screenshot", screenshotPath] });
-			assert.equal(screenshot.isError, false, JSON.stringify(screenshot));
-			assert.equal((await stat(screenshotPath)).isFile(), true);
-
-			const close = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["close"] });
-			assert.equal(close.isError, false, JSON.stringify(close));
-
-			const invocations = await readInvocationLog(logPath);
-			assert.equal(invocations.filter((entry) => entry.args.includes("close")).length, 1);
-		});
-	} finally {
-		await rm(tempDir, { force: true, recursive: true });
-	}
-});
-
-test("agentBrowserExtension resolves relative prompt screenshot paths before allowing close", { concurrency: false }, async () => {
-	const tempDir = await mkdtemp(join(tmpdir(), "pi-agent-browser-relative-artifact-"));
-	const logPath = join(tempDir, "invocations.log");
-	const relativeScreenshotPath = "./release-smoke.png";
-	const absoluteScreenshotPath = join(tempDir, "release-smoke.png");
-	const basePath = process.env.PATH ?? "";
-	await writeFakeAgentBrowserBinary(
-		tempDir,
-		`const fs = require("node:fs");
-const args = process.argv.slice(2);
-fs.appendFileSync(${JSON.stringify(logPath)}, JSON.stringify({ args }) + "\\n");
-if (args.includes("screenshot")) {
-  const path = args[args.length - 1];
-  fs.writeFileSync(path, "png");
-  process.stdout.write(JSON.stringify({ success: true, data: { path } }));
-} else if (args.includes("close")) {
-  process.stdout.write(JSON.stringify({ success: true, data: { closed: true } }));
-} else {
-  process.stdout.write(JSON.stringify({ success: true, data: "ok" }));
-}`,
-	);
-
-	try {
-		await withPatchedEnv({ PATH: `${tempDir}:${basePath}` }, async () => {
-			const harness = createExtensionHarness({ cwd: tempDir, prompt: `Save a screenshot here: ${relativeScreenshotPath}` });
-			await runExtensionEvent(harness.handlers, "session_start", { reason: "new" }, harness.ctx);
-
-			const blockedClose = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["close"] });
-			assert.equal(blockedClose.isError, true);
-			assert.equal((blockedClose.details?.promptGuard as { missingArtifacts?: Array<{ path?: string }> } | undefined)?.missingArtifacts?.[0]?.path, relativeScreenshotPath);
-
-			const screenshot = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["screenshot", relativeScreenshotPath] });
-			assert.equal(screenshot.isError, false, JSON.stringify(screenshot));
-			assert.equal((await stat(absoluteScreenshotPath)).isFile(), true);
-
-			const close = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["close"] });
-			assert.equal(close.isError, false, JSON.stringify(close));
-
-			const invocations = await readInvocationLog(logPath);
-			assert.equal(invocations.filter((entry) => entry.args.includes("close")).length, 1);
-		});
-	} finally {
-		await rm(tempDir, { force: true, recursive: true });
-	}
-});
-
-test("agentBrowserExtension reports click dispatch diagnostic when upstream reports success without dispatching DOM events", { concurrency: false }, async () => {
-	const tempDir = await mkdtemp(join(tmpdir(), "pi-agent-browser-click-dispatch-"));
-	const logPath = join(tempDir, "invocations.log");
-	const basePath = process.env.PATH ?? "";
-	await writeFakeAgentBrowserBinary(
-		tempDir,
-		`const fs = require("node:fs");
-const args = process.argv.slice(2);
-const stdin = fs.readFileSync(0, "utf8");
-fs.appendFileSync(${JSON.stringify(logPath)}, JSON.stringify({ args, stdin }) + "\\n");
-if (args.includes("snapshot")) {
-  process.stdout.write(JSON.stringify({ success: true, data: {
-    origin: "https://shop.example/inventory",
-    refs: {
-      e1: { role: "button", name: "Add to cart" },
-      e2: { role: "button", name: "Add to cart" }
-    },
-    snapshot: '- button "Add to cart" [ref=e1]\\n- button "Add to cart" [ref=e2]'
-  } }));
-} else if (args.includes("click")) {
-  process.stdout.write(JSON.stringify({ success: true, data: { clicked: args[args.length - 1] } }));
-} else if (args.includes("eval")) {
-  if (stdin.includes("window[marker] = state")) {
-    process.stdout.write(JSON.stringify({ success: true, data: { result: { status: "installed" } } }));
-  } else if (stdin.includes("no-native-event-observed")) {
-    process.stdout.write(JSON.stringify({ success: true, data: { result: { status: "no-native-event-observed", nativeEventCount: 0 } } }));
-  } else {
-    process.stdout.write(JSON.stringify({ success: true, data: { result: { title: "Shop", url: "https://shop.example/inventory" } } }));
-  }
-} else {
-  process.stdout.write(JSON.stringify({ success: true, data: "ok" }));
-}`,
-	);
-
-	try {
-		await withPatchedEnv({ PATH: `${tempDir}:${basePath}` }, async () => {
-			const harness = createExtensionHarness({ cwd: tempDir });
-			await runExtensionEvent(harness.handlers, "session_start", { reason: "new" }, harness.ctx);
-
-			const click = await executeRegisteredTool(harness.tool, harness.ctx, { args: ["click", "[data-test=add-to-cart]"] });
-			assert.equal(click.isError, true);
-			assert.match((click.content[0] as { text: string }).text, /Click dispatch diagnostic:/);
-			assert.equal((click.details?.clickDispatch as { status?: string } | undefined)?.status, "no-native-event-observed");
-			assert.equal((click.details?.clickDispatch as { target?: { kind?: string; selector?: string } } | undefined)?.target?.kind, "selector");
-			assert.equal((click.details?.clickDispatch as { target?: { kind?: string; selector?: string } } | undefined)?.target?.selector, "[data-test=add-to-cart]");
-			assert.ok((click.details?.nextActions as Array<{ id?: string }> | undefined)?.some((action) => action.id === "retry-click-after-dispatch-miss"));
-
-			const invocations = await readInvocationLog(logPath);
-			assert.equal(invocations.filter((entry) => entry.args.includes("click")).length, 1);
-			assert.ok(invocations.some((entry) => entry.args.includes("eval") && (entry.stdin ?? "").includes("window[marker] = state")));
-			assert.ok(invocations.some((entry) => entry.args.includes("eval") && (entry.stdin ?? "").includes("no-native-event-observed")));
 		});
 	} finally {
 		await rm(tempDir, { force: true, recursive: true });
